% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{actions}
\alias{actions}
\alias{Actions_Carrier_Drug}
\alias{Actions_Enzyme_Drug}
\alias{Actions_Target_Drug}
\alias{Actions_Transporter_Drug}
\title{Drug Carriers/ Enzymes/ Targets/ Transporters related Actions}
\format{
a tibble with 2 variables:
\describe{
  \item{text}{describe related action}
  \item{\emph{parent_id}}{drug/ carrier/ target/ enzyme/ transporter id}
}

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 109 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 653 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3394 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 338 rows and 2 columns.
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank
Documentation}
}
\usage{
Actions_Carrier_Drug

Actions_Enzyme_Drug

Actions_Target_Drug

Actions_Transporter_Drug
}
\description{
A collection of actions related to drugs carriers
}
\keyword{datasets}
