% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Affected_Organisms_Drug}
\alias{Affected_Organisms_Drug}
\title{Drug related Affected Organisms}
\format{
a tibble with 2 variables:
\describe{
  \item{text}{affected organism description}
  \item{parent_id}{drugbank id}
}
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Affected_Organisms_Drug
}
\description{
A list of organisms in which the drug may display activity; activity may
depend on local susceptibility patterns and resistance.
}
\keyword{datasets}
