% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{articles}
\alias{articles}
\alias{Articles_Carrier_Drug}
\alias{Articles_Drug}
\alias{Articles_Enzyme_Drug}
\alias{Articles_Target_Drug}
\alias{Articles_Transporter_Drug}
\title{Drugs/ Carriers/ Enzymes/ Targets/ Transporters related Articles}
\format{
a tibble with 4 variables:
\describe{
  \item{ref-id}{Identifier for the article being referenced.
  This is unique across all reference types (books, links, article).}
  \item{pubmed-id}{The PubMed identifier for the article.}
  \item{citation}{Article citation in a standard format.}
  \item{\emph{parent_id}}{drug/carrier/target/enzyme/transporter id}
}

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 635 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 13139 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5595 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 21775 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2231 rows and 4 columns.
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Articles_Carrier_Drug

Articles_Drug

Articles_Enzyme_Drug

Articles_Target_Drug

Articles_Transporter_Drug
}
\description{
A list of articles that were used as references for drugs carriers
}
\keyword{datasets}
