% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attachments}
\alias{attachments}
\alias{Attachments}
\alias{Attachments_Enzymes}
\alias{Attachments_Carriers}
\alias{Attachments_Targets}
\alias{Attachments_Transporters}
\title{Drugs/ Carriers/ Enzymes/ Targets/ Transporters Attachments}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 424 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 413 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 101 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 67 rows and 4 columns.
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Attachments

Attachments_Enzymes

Attachments_Carriers

Attachments_Targets

Attachments_Transporters
}
\value{
a tibble with 4 variables:
\describe{
  \item{ref-id}{Identifier for the article being referenced.
  This is unique across all reference types (books, links, article,
  attachments).}
  \item{title}{The title of the attachment.}
  \item{url}{The url to download the attachment from.}
  \item{\emph{parent_id}}{drug/carrier/target/enzyme/transporter id}
}
}
\description{
Return a list of attachment that were used as references for drugs or CETT
}
\keyword{datasets}
