% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{books}
\alias{books}
\alias{Books_Drug}
\alias{Textbooks_Carrier_Drug}
\alias{Textbooks_Enzyme_Drug}
\alias{Textbooks_Target_Drug}
\alias{Textbooks_Transporter_Drug}
\title{Drugs/ Carriers/ Enzymes/ Targets/ Transporters related Books}
\format{
a tibble with 4 variables:
\describe{
  \item{ref-id}{Identifier for the article being referenced.
  This is unique across all reference types (books, links, article).}
  \item{isbn}{ISBN identifying the textbook.}
  \item{citation}{A Textbook citation in a standard format.}
  \item{\emph{parent_id}}{drug/ carrier/ target/ enzyme/ transporter id}
}

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 434 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 21 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 78 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 107 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 12 rows and 4 columns.
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Books_Drug

Textbooks_Carrier_Drug

Textbooks_Enzyme_Drug

Textbooks_Target_Drug

Textbooks_Transporter_Drug
}
\description{
A list of text books that were used as references for drugs
}
\keyword{datasets}
