% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Categories_Drug}
\alias{Categories_Drug}
\title{Drug Categories
General categorizations of the drug}
\format{
a tibble with 3 variables:
\describe{
  \item{category}{Category name}
  \item{mesh-id}{\href{https://www.nlm.nih.gov/mesh/}{The Medical Subjects
  Headings (MeSH)} identifier for the category.}
  \item{parent_id}{drugbank id}
}
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Categories_Drug
}
\description{
Each drug may have one or more category.
}
\keyword{datasets}
