% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cett}
\alias{cett}
\alias{Carriers_Drug}
\alias{Enzymes_Drug}
\alias{Targets_Drug}
\alias{Transporters_Drug}
\title{Carriers/ Enzymes/ Targets/ Transporters}
\format{
a tibble with 6 variables:
\describe{
  \item{id}{Identifier for the record}
  \item{name}{related name}
  \item{organism}{Organism that the protein comes from.}
  \item{known_action}{Whether the pharmacological action of the drug is due
   to this target interaction.}
  \item{position}{related position}
  \item{parent_id}{drugbank id}
}

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 805 rows and 6 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5158 rows and 8 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 18780 rows and 6 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3050 rows and 6 columns.
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Carriers_Drug

Enzymes_Drug

Targets_Drug

Transporters_Drug
}
\description{
Protein targets of drug action, enzymes that are inhibited/induced or
involved in metabolism, and carrier or transporter proteins involved in
movement of the drug across biological membranes.
}
\details{
Each of targets, enzymes, carriers and transporters contain one or more
child elements tibbles
}
\keyword{datasets}
