% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dosages_Drug}
\alias{Dosages_Drug}
\title{Drug Dosages
A list of the commercially available dosages of the drug.}
\format{
a tibble with 4 variables:
\describe{
  \item{form}{The pharmaceutical formulation by which the drug is introduced
   into the body.}
  \item{route}{The path by which the drug or product is taken into the body}
  \item{strength}{The amount of active drug ingredient provided in the
  dosage}
  \item{parent_id}{drugbank id}
}
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Dosages_Drug
}
\description{
Each drug may have one or more dosage.
}
\keyword{datasets}
