% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drugbank}
\alias{drugbank}
\title{DrugBank Database}
\format{
An object of class \code{dvobject} of length 5.
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank
Documentation}
}
\usage{
drugbank
}
\description{
dvobject containing list of dataframes of parsed DrugBank database
}
\keyword{datasets}
