% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Enzymes_Reactions_Drug}
\alias{Enzymes_Reactions_Drug}
\title{Drug Reactions Enzymes
Enzymes involved in metabolizing this drug.}
\format{
a tibble with 3 variables:
\describe{
  \item{drugbank-id}{}
  \item{name}{}
  \item{uniprot-id}{uniprot id}
}
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Enzymes_Reactions_Drug
}
\description{
Drug Reactions Enzymes
Enzymes involved in metabolizing this drug.
}
\keyword{datasets}
