% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Food_Interactions_Drug}
\alias{Food_Interactions_Drug}
\title{Drug Food Interactions}
\format{
a tibble 2 variables:
\describe{
  \item{interaction}{descripts of interactions this drug has with food.}
  \item{parent_key}{drugbank key}
}
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Food_Interactions_Drug
}
\description{
Description of interactions this drug has with food.
}
\details{
Each drug may have one or more food interaction.
}
\keyword{datasets}
