% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go}
\alias{go}
\alias{GO_Classifiers_Polypeptide_Carrier_Drug}
\alias{GO_Classifiers_Polypeptides_Enzyme_Drug}
\alias{GO_Classifiers_Polypeptide_Target_Drug}
\alias{GO_Classifiers_Polypeptide_Transporter_Drug}
\title{Carriers/ Enzymes/ Targets/ Transporters related Gene Ontology (GO)}
\format{
a tibble with 3 variables:
\describe{
  \item{category}{}
  \item{description}{}
  \item{polypeptide_id}{polypeptide id}
}

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1832 rows and 3 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8495 rows and 3 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 79214 rows and 3 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4752 rows and 3 columns.
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
GO_Classifiers_Polypeptide_Carrier_Drug

GO_Classifiers_Polypeptides_Enzyme_Drug

GO_Classifiers_Polypeptide_Target_Drug

GO_Classifiers_Polypeptide_Transporter_Drug
}
\description{
The \href{http://geneontology.org/}{Gene Ontology (GO)} Consortium identifier
}
\details{
Each Carriers/ Enzymes/ Targets/ may have zero or more GO.
}
\keyword{datasets}
