% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Groups_Drug}
\alias{Groups_Drug}
\title{Drug Groups}
\format{
a tibble with 2 variables:
\describe{
  \item{group}{one of the above values}
  \item{drugbank-id}{drugbank id}
}
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Groups_Drug
}
\description{
Groups that this drug belongs to.
}
\details{
Each drug may have one or more category.

Groups include:
\itemize{
 \item approved
 \item vet_approved
 \item nutraceutical
 \item illicit
 \item withdrawn
 \item investigational
 \item experimental
 }
}
\keyword{datasets}
