% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{International_Brands_Drug}
\alias{International_Brands_Drug}
\title{Drug International Brands}
\format{
a tibble with 3 variables:
\describe{
  \item{name}{The proprietary, well-known name for given to this drug by a
  manufacturer.}
  \item{company}{The company or manufacturer that uses this name.}
  \item{parent_key}{drugbank id}
}
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
International_Brands_Drug
}
\description{
The proprietary names used by the manufacturers for commercially available
forms of the drug, focusing on brand names for products that are available
in countries other than Canada and the Unites States.
}
\details{
Each drug may have one or more international brand.
}
\keyword{datasets}
