% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{links}
\alias{links}
\alias{Links_Drug}
\alias{Links_Carrier_Drug}
\alias{Links_Enzyme_Drug}
\alias{Links_Target_Drug}
\alias{Links_Transporter_Drug}
\title{Drugs/ Carriers/ Enzymes/ Targets/ Transporters related Links}
\format{
a tibble with 3 variables:
\describe{
  \item{ref-id}{Name of the source website}
  \item{title}{Identifier for this drug in the given resource}
  \item{url}{The url of the website}
  \item{\emph{parent_id}}{drug/ carrier/ target/ enzyme/ transporter id}
}

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4300 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 110 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 848 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 691 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 266 rows and 4 columns.
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Links_Drug

Links_Carrier_Drug

Links_Enzyme_Drug

Links_Target_Drug

Links_Transporter_Drug
}
\description{
A list of websites that were used as references for
Drugs/ Carriers/ Enzymes/ Targets/ Transporters
}
\details{
Each drug/ carrier/ target/ enzyme/ transporter may have one or more link.
}
\keyword{datasets}
