% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Manufacturers_Drug}
\alias{Manufacturers_Drug}
\title{Drug Manufacturers}
\format{
a tibble with 3 variables:
\describe{
  \item{text}{the name or description of the manufacturer}
  \item{parent_key}{drugbank id}
}
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Manufacturers_Drug
}
\description{
A list of companies that are manufacturing the commercially available forms
of this drug that are available in Canada and the Unites States.
}
\details{
Each drug may have one or more Manufacturer.
}
\keyword{datasets}
