% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mixtures_Drug}
\alias{Mixtures_Drug}
\title{Drug Mixture}
\format{
a tibble with 3 variables:
\describe{
  \item{name}{The proprietary name provided by the manufacturer for this
   combination product.}
  \item{ingredients}{A list of ingredients, separated by addition symbols.}
  \item{parent_key}{drugbank id}
}
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Mixtures_Drug
}
\description{
All commercially available products in which this drug is available in
 combination with other drug molecules.
}
\details{
Each drug may have one or more mixture.
}
\keyword{datasets}
