% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Patents_Drug}
\alias{Patents_Drug}
\title{Drug Patent}
\format{
a tibble with 6 variables:
\describe{
  \item{number}{The patent number(s) associated with the drug}
  \item{country}{The country that issued the patent rights}
  \item{approved}{The date that the patent request was filed}
  \item{expires}{The date that the patent rights expire}
  \item{pediatric-extension}{Indicates whether or not a pediatric extension
  has been approved for the patent. Granted pediatric extensions provide an
  additional 6 months of market protection}
  \item{parent_key}{drugbank id}
}
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Patents_Drug
}
\description{
A property right issued by the \href{http://www.uspto.gov/}{United States
Patent and Trademark Office (USPTO)} to an inventor for a limited time,
 in exchange for public disclosure of the invention when the patent is
 granted. Drugs may be issued multiple patents.
}
\details{
Each drug may have one or more patent.
}
\keyword{datasets}
