% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pathways_Drug}
\alias{Pathways_Drug}
\title{Drug Pathways}
\format{
a tibble with 4 variables:
\describe{
  \item{smpdb_id}{\href{http://smpdb.ca/}{Small Molecule Protein Database
 (SMPDB)} identifier for this pathway.}
  \item{name}{Pathway name}
  \item{category}{Pathway category}
  \item{parent_key}{drugbank id}
}
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Pathways_Drug
}
\description{
Metabolic, disease, and biological pathways that the drug is involved in,
as identified by the \href{http://smpdb.ca/}{Small Molecule Protein Database
 (SMPDB)}.
}
\details{
Each drug may have one or more pathway.
}
\keyword{datasets}
