% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PDB_Entries_Drug}
\alias{PDB_Entries_Drug}
\title{Drug PDB Entries}
\format{
a tibble with 2 variables:
\describe{
  \item{text}{PDB identifier}
  \item{parent_key}{drugbank id}
}
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
PDB_Entries_Drug
}
\description{
Protein Data Bank (PDB) identifiers for this drug
}
\details{
Each drug may have one or more PDB Entry
}
\keyword{datasets}
