% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfam}
\alias{pfam}
\alias{PFAMS_Polypeptide_Carrier_Drug}
\alias{PFAMS_Polypeptides_Enzyme_Drug}
\alias{PFAMS_Polypeptide_Target_Drug}
\alias{PFAMS_Polypeptid_Transporter_Drug}
\title{PFAMs}
\format{
a tibble with 3 variables:
\describe{
  \item{identifier}{}
  \item{name}{}
  \item{polypeptide_id}{polypeptide id}
}

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 133 rows and 3 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 705 rows and 3 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8328 rows and 3 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 409 rows and 3 columns.
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
PFAMS_Polypeptide_Carrier_Drug

PFAMS_Polypeptides_Enzyme_Drug

PFAMS_Polypeptide_Target_Drug

PFAMS_Polypeptid_Transporter_Drug
}
\description{
The \href{http://pfam.xfam.org/}{protein family (pfam)} identifier
}
\details{
Each Polypeptid may have one or more PFAM.
}
\keyword{datasets}
