% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poly_syn}
\alias{poly_syn}
\alias{Synonyms_Polypeptide_Carrier_Drug}
\alias{Synonyms_Polypeptides_Enzyme_Drug}
\alias{Synonyms_Polypeptide_Target_Drug}
\alias{Synonyms_Polypeptide_Transporter_Drug}
\title{Polypeptide Synonyms}
\format{
a tibble with 2 variables:
\describe{
  \item{syn}{alternative name}
  \item{polypeptide_id}{polypeptide id}
}

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 332 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2353 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 17979 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1323 rows and 2 columns.
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Synonyms_Polypeptide_Carrier_Drug

Synonyms_Polypeptides_Enzyme_Drug

Synonyms_Polypeptide_Target_Drug

Synonyms_Polypeptide_Transporter_Drug
}
\description{
Alternate names or identifiers that may be associated with this polypeptide
}
\details{
Each element may have one or more synonyms.
}
\keyword{datasets}
