% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Prices_Drug}
\alias{Prices_Drug}
\title{Drug Prices}
\format{
a tibble with 5 variables:
\describe{
  \item{description}{}
  \item{cost}{Drug price per unit}
  \item{unit}{}
  \item{currency}{Currency of price, example: USD}
  \item{parent_key}{drugbank id}
}
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Prices_Drug
}
\description{
Unit drug prices
}
\details{
Each drug may have one or more price
}
\keyword{datasets}
