% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Reactions_Drug}
\alias{Reactions_Drug}
\title{Drug Reactions}
\format{
a tibble with 6 variables:
\describe{
  \item{sequence}{Reactions are displayed within a numerical sequence.}
  \item{left_drugbank_id}{The substrate of the reaction. May be a drug or a
  metabolite.}
  \item{left_drugbank_name}{}
  \item{right_drugbank_id}{The product of the reaction. May be a drug or a
   metabolite.}
  \item{right_drugbank_name}{}
  \item{parent_key}{drugbank id}
}
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Reactions_Drug
}
\description{
A sequential representation of the metabolic reactions that this drug
molecule is involved in. Depending on available information, this may include
metabolizing enzymes, reaction type, substrates, products, pharmacological
activity of metabolites, and a structural representation of the biochemical
reactions.
}
\details{
Each drug may have one or more reaction.
}
\keyword{datasets}
