% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sequences_Drug}
\alias{Sequences_Drug}
\title{Drug Sequences}
\format{
a tibble with 3 variables:
\describe{
  \item{sequence}{}
  \item{format}{}
  \item{parent_key}{drugbank id}
}
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
Sequences_Drug
}
\description{
The amino acid sequence; provided if the drug is a peptide.
}
\details{
Each drug may have one or more sequence.

Describes peptide sequences of biotech drugs. The sequence variable contains
a textual representation of the sequence, in the format described by the
format variable. Currently, only the \strong{FASTA} format is used.
}
\keyword{datasets}
