% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SNP_Adverse_Drug_Reactions_Drug}
\alias{SNP_Adverse_Drug_Reactions_Drug}
\title{SNP Adverse Drug Reactions}
\format{
a tibble with 9 variables:
\describe{
  \item{protein-name}{Proteins involved in this SNP.}
  \item{gene-symbol}{Genes involved in this SNP.}
  \item{uniprot-id}{\href{http://www.uniprot.org/}{Universal Protein
  Resource (UniProt)} identifiers for proteins involved in this pathway.}
  \item{rs-id}{The \href{https://www.ncbi.nlm.nih.gov/projects/SNP/}{SNP
  Database} identifier for this single nucleotide polymorphism.}
  \item{allele}{The alleles associated with the identified SNP.}
  \item{adverse-reaction}{}
  \item{description}{}
  \item{pubmed-id}{Reference to PubMed article.}
  \item{parent_key}{\href{http://www.ncbi.nlm.nih.gov/pubmed}{drugbank id}}
}
}
\source{
\href{https://docs.drugbankplus.com/xml/}{Drugbank Documentation}
}
\usage{
SNP_Adverse_Drug_Reactions_Drug
}
\description{
The adverse drug reactions that may occur as a result of the listed single
nucleotide polymorphisms (SNPs).
}
\details{
Each drug may have one or more SNP Adverse.
}
\keyword{datasets}
